Cleaned up reservebot.go, moved LookupEnv funcs into utils, created parseAdmin func and moved into utils
