+++ util/util.go
@@ -2,6 +2,7 @@ package util
 
 import (
 	"math"
+	"os"
 	"strconv"
 )
 
@@ -30,4 +31,44 @@ func Ordinalize(num int) string {
 
 	return strconv.Itoa(num) + ordinalDictionary[positiveNum]
 
+}
+
+func LookupEnvOrString(key string, defaultVal string) string {
+        if val, ok := os.LookupEnv(key); ok {
+                return val
+        }
+        return defaultVal
+}
+
+func LookupEnvOrInt(key string, defaultVal int) int {
+        if val, ok := os.LookupEnv(key); ok {
+                v, _ := strconv.Atoi(val)
+                return v
+        }
+        return defaultVal
+}
+
+func LookupEnvOrBool(key string, defaultVal bool) bool {
+        if val, ok := os.LookupEnv(key); ok {
+                if val == "true" {
+                        return true
+                } else {
+                        return false
+                }
+        }
+        return defaultVal
+}
+
+func parseAdmins(admins string) []string {
+        // Convert admins list into slice
+        var admins_ary []string
+        if len(admins) > 0 {
+                if strings.Contains(admins, ",") {
+                        admins_ary = strings.Split(admins, ",")
+                } else {
+                        admins_ary = append(admins_ary, admins)
+                }
+        }
+
+	return admins_ary
 }
